### 2023-04-02 #5323 @editorialbot
[PRE REVIEW]: GPCERF - An R package for implementing Gaussian processes for estimating causal exposure response curves
**Submitting author:** @naeemkh (<a href="http://orcid.org/0000-0003-4315-1426">Naeem Khoshnevis</a>)
**Repository:** https://github.com/NSAPH-Software/GPCERF
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.2.1
