### 2023-04-02 #5323 @spholmes
Hi @thomaspinder, could you review this submission it does gaussian processes and bayesian inference, I thought you might be interested. Thanks.
