### 2023-05-01 #5323 @martinmodrak
Looks cool, I'll review.

Dne po 1. 5. 2023 15:57 uživatel Susan Holmes ***@***.***>
napsal:

> 👋 @martinmodrak <https://github.com/martinmodrak> would it be possible
> for you to review the GPCERF package for JOSS, many thanks, Susan
>
