### 2023-05-26 #5465 @nhejazi
@spholmes -- Apologies for the late realization of this, but @Naeemkh has brought to my attention that our mutual affiliation with Harvard University disqualifies me as a reviewer on account of JOSS's COI policy. In my initial reading, I'd missed the last line of the COI policy, leading to the working assumption that my service in this capacity would be acceptable since we do not and have not interacted at all outside of the context of this review.
