### 2023-06-02 #5465 @Naeemkh
Hi @spholmes, I would appreciate if you could guide us regarding the next course of action for this submission. Is it possible to waive @nhejazi 's Conflict of Interest (COI) and allow him to proceed with the review process? Please inform us at your earliest convenience.
