### 2023-06-04 #5465 @spholmes
> Hi @spholmes, I would appreciate if you could guide us regarding the next course of action for this submission. Is it possible to waive @nhejazi 's Conflict of Interest (COI) and allow him to proceed with the review process? Please inform us at your earliest convenience.

Sorry @Naeemkh , we can't waive the conflict in this case because @nhejazi  and you work at the same institution. If you want to suggest some other reviewers with whom you have no connections or COI, I can ask them, I will also use the data base tomorrow to try and fill in for the next round.
