### 2023-06-06 #5323 @spholmes
Hi @andrewherren, would you be available to look at the submission about the GPCERF package for the JOSS.
