### 2023-06-13 #5323 @spholmes
👋 @andrewherren : would you be willing and available to review this submission to JOSS?
Many Thanks
Susan
