### 2023-07-24 NSAPH-Software/GPCERF#66 @martinmodrak
The example in README.md requires the ranger library, but this is not mentioned
This is a minor thing, but following the installation instructions and running the demo, I get a bunch of errors like:

```
Loading required namespace: ranger
Failed with error:  ‘there is no package called ‘ranger’’
Error : loading required package (ranger) failed
Loading required namespace: ranger
Failed with error:  ‘there is no package called ‘ranger’’
In addition: Warning message:
In FUN(X[[i]], ...) : Error in algorithm m_ranger
  The Algorithm will be removed from the Super Learner (i.e. given weight 0)
```

The example works even without `ranger`, but it probably should instruct the users to install the `ranger` library (or avoid using it).
