### 2023-07-24 NSAPH-Software/GPCERF#67 @martinmodrak
Mild mismatch between DESCRIPTION file, commit history and JOSS authors list
This might actually not be a real issue and just need a little bit of explanation to resolve.

The DESCRIPTION file has Tanujit, Dey as a contributor to the package, but they are not listed as co-authors of the JOSS paper.

Tanujit, Dey and Danielle, Braun are listed in DESCRIPTION but have made no contributions to the repository.

Could you briefly clarify the role of those two contributors in the project and why they are/aren't included in the authors list?
