### 2023-07-24 NSAPH-Software/GPCERF#68 @martinmodrak
JOSS paper should at least briefly discuss other GP packages
The paper currently discusses only R packages for causal inference, but avoids relating GPCERF to other packages for Gaussian process inference.

It appears to me that GPCERF performs several steps: 1) searching for hyperparameters, 2) computing the generalized propensity score (GPS) 3) fitting a 2D Gaussian process for the outcome using the exposure and propensity score.

It seems that at least in principle, any package capable of 2D GP processing can be used for step 3. If that's correct, it IMHO should be at least briefly discussed and explained how the GP implementation in this package relates to other common GP packages. Especially since some packages would allow for estimating the alpha, beta and gamma hyperparameters jointly with the GP, which would IMHO be a superior approach (if not, a brief explanation why is IMHO warranted).

It is also possible I am missing something about the nature of the computations in GPCERF, so feel free to correct any misunderstanding.
