### 2023-07-24 NSAPH-Software/GPCERF#69 @martinmodrak
Missing information on bug reports and where to get help
JOSS requirements include:
"Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

Item 1) is well covered, but unless I am missing something, the other two seem to have been neglected. Note that this doesn't have to be anything complex - just stating in the README.md that issues should be reported on GitHub + providing an issue template should satisfy 2) and having "contact XX at EMAIL for additional support" should satisfy 3)
