### 2023-07-24 NSAPH-Software/GPCERF#71 @martinmodrak
The GP can be unreasonably overconfident
It is possible I didn't use the package correctly, but a relatively direct modification of the examples yields reliably wrong, overconfident results.

Here's the code I am using:

```r
library(GPCERF)
library(ggplot2)
set.seed(32236)
N <- 500
use_nn <- FALSE

sim_data <- data.frame(Y = rep(NA_real_, N), treat = rnorm(N), cf1 = rnorm(N), cf2 = rnorm(N))
sim_data$Y <- rnorm(N, mean = sin(2 * sim_data$treat) + cos(2 * sim_data$cf1))

# exposure values
q1 <- stats::quantile(sim_data$treat, 0.05)
q2 <- stats::quantile(sim_data$treat, 0.95)

true_df <- data.frame(w = seq(q1, q2, length.out = 100))
true_df$y = sin(2 * true_df$w)

n_core <- 20

m_xgboost <- function(nthread = n_core, ...) {
  SuperLearner::SL.xgboost(nthread = nthread, ...)
}

m_ranger <- function(num.threads = n_core, ...){
  SuperLearner::SL.ranger(num.threads = num.threads, ...)
}

# Estimate GPS function
gps_m <- estimate_gps(cov_mt = sim_data[,-(1:2)],
                      w_all = sim_data$treat,
                      sl_lib = c("m_xgboost", "m_ranger"),
                      dnorm_log = TRUE)

w_all <- seq(q1, q2, length.out = 100)

params_lst <- list(alpha = 10 ^ seq(-2, 1/2, length.out = 10),
                   beta = 10 ^ seq(-2, 1/2, length.out = 10),
                   g_sigma = c(0.1, 0.5, 1, 5, 10),
                   tune_app = "all")

if(use_nn) {
  params_lst$n_neighbor <- 50
  params_lst$block_size <-  1e3
  cerf_gp_obj <- estimate_cerf_nngp(sim_data,
                                    w_all,
                                    gps_m,
                                    params = params_lst,
                                    nthread = 12)
} else {
  cerf_gp_obj <- estimate_cerf_gp(sim_data,
                                  w_all,
                                  gps_m,
                                  params = params_lst,
                                  nthread = n_core)
}

plot_fun <- function(object, ...) {
    # extract data
  tmp_data <- data.frame(w_vals = object$posterior$w,
                         mean_vals = object$posterior$mean,
                         sd_vals = object$posterior$sd)

  g1 <- ggplot2::ggplot(tmp_data) +
       ggplot2::geom_ribbon(ggplot2::aes(.data$w_vals,
                                y = .data$mean_vals,
                                ymin = .data$mean_vals - 1.96 * .data$sd_vals,
                                ymax = .data$mean_vals + 1.96 * .data$sd_vals),
                                fill = "blue", alpha = 0.25) +
        ggplot2::geom_line(ggplot2::aes(.data$w_vals, .data$mean_vals),
                           color = "blue", size = 1) +
        ggplot2::theme_bw() +
        ggplot2::ggtitle("Estimated CERF (gp) with credible band (1.96sd)") +
        ggplot2::xlab("Exposure level") +
        ggplot2::ylab("Population average counterfactual outcome")

  return(g1)
}

plot_fun(cerf_gp_obj) +
  geom_line(data = true_df,
            aes(x = w, y = y, color = "True"), size = 1.5)

```

Which results in:

![image](https://github.com/NSAPH-Software/GPCERF/assets/9483603/9825800a-a987-4e21-a118-bd50a41fa207)

I.e. the qualitative trend is there, but most of the true response curve is outside the uncertainty interval.

The situation does not improve when using more data and the nnGP - replace the config lines with

```r
N <- 5000
use_nn <- TRUE
```

and you get:

![image](https://github.com/NSAPH-Software/GPCERF/assets/9483603/b7c173bf-e037-428c-9890-5ee790218a81)

Am I using the package wrong? It seems to me that the hyperparameter grid contains parameters that should fit the data well... Or is there a reason to expect poor performance for this type of relationship?
