### 2023-07-24 #5465 @martinmodrak
Sorry for the delay. I have just completed a first round of reviewing the package. Overall, this is a nice submission and a useful package.

I've reported one issue possible suggesting problems with the computation (https://github.com/NSAPH-Software/GPCERF/issues/71), but it might as well be me misunderstanding the package - resolving/explaining this problem is in my view important for acceptance.

Beyond that I've noticed some minor issues which I have reported in the repository, nothing of this is critical and I am open to authors arguing that those are not necessary for acceptance.
