### 2023-07-24 NSAPH-Software/GPCERF#72 @martinmodrak
Guide to choosing hyperparameter grid
Choice if reasonable range of values for the hyperparameter candidates in the grid search seems to be important for good inferences. I would expect at least some basic guidance on the topic to appear somewhere in the documentation (e.g. that there's no use for length scales that are shorter than minimum distance between the data points or comparable to the range of the data).
