### 2023-07-24 NSAPH-Software/GPCERF#71 @martinmodrak
An interesting variant of the problem arises with:

```r
N <- 500
use_nn <- FALSE

sim_data <- data.frame(Y = rep(NA_real_, N), treat = rnorm(N), cf1 = rnorm(N), cf2 = rnorm(N))
sim_data$Y <- rnorm(N, mean = cos(2 * sim_data$cf1))

# exposure values
q1 <- stats::quantile(sim_data$treat, 0.05)
q2 <- stats::quantile(sim_data$treat, 0.95)

true_df <- data.frame(w = seq(q1, q2, length.out = 100))
true_df$y = 0
```

Now the result correctly concentrate on flat response curve, but the curve is shifted away from 0:

![image](https://github.com/NSAPH-Software/GPCERF/assets/9483603/a23ea896-6095-43d4-8fe6-a972b0c809a6)
