### 2023-08-08 NSAPH-Software/GPCERF#70 @Naeemkh
Thanks @martinmodrak, this really improved the code.
