### 2023-10-03 NSAPH-Software/GPCERF#67 @Naeemkh
Thank you for bringing up the discrepancies in the authorship and contributor list. We appreciate the chance to clarify the roles of both individuals.
Danielle Braun played a pivotal role during the conceptualization phase and provided key insights during meetings and discussions on how to develop the software and its underlying scientific principles. While her contributions may not be evident through direct commits or
code contributions in the repository, her ideas, feedback, and domain expertise significantly shaped the development of the package. Given the importance of her role in the foundational stages and her sustained involvement throughout, she was deemed deserving of authorship status in the JOSS paper.
Tanujit Dey was initially scheduled to contribute through testing and other implementation-related activities. Given other commitments and opportunities he prioritized, he could not have time to complete the tasks as initially envisioned. Since his contributions were minimal and did not shape the final product, he was listed as a contributor in the DESCRIPTION file of the package but was not included as a co-author in the JOSS paper.
We hope this clarifies any concerns. We are open to any further feedback or questions you may have on this matter.
