### 2023-10-03 NSAPH-Software/GPCERF#68 @Naeemkh
Thanks @martinmodrak, You are correct. There are other R packages that perform Gaussian process regression. The main reason we chose to not directly use these packages is that all of them utilize traditional hyper-parameter tuning approaches, either by sampling from the posterior distributions of the hyper-parameters or by maximizing the marginal likelihood function, while the tuning approach we proposed in the paper is based on maximizing covariate balancing. We updated the paper to reflect this discussion.
