### 2023-10-04 #5465 @Kevin-Mattheus-Moerman
@spholmes in order to add `@andrewherren` as a reviewer you need to call `@editorialbot add @andrewherren as reviewer`. Once that is complete they can generate their checklist using: `@editorialbot generate my checklist`
