### 2023-10-04 NSAPH-Software/GPCERF#71 @boyuren158
We thank the reviewer for raising these two special cases where our package appears to be overconfident. These two examples motivate a more refined strategy for the specification of hyper-parameter grids, which we will elaborate in the response to another [issue](https://github.com/NSAPH-Software/GPCERF/issues/72).

We would like to first clarify that the way the reviewer calculates the true CERF is incorrect. We know that in the first example the counterfactual outcome is generated by the following model $$E[Y(w)|X_1,X_2] = \sin(2w) + \cos(2X_1).$$ The true CERF is then $$E[Y(w)] = \sin(2w) + \mathbb E(\cos(2X_1)) = \sin(2w) + 0.133.$$ The last equality is due to
$$X_1\sim N(0,1).$$ The same calculation applies for the second example, and the true CERF there is
$$E[Y(w)] = \mathbb E(\cos(2X_1)) = 0.133.$$ A constant shift 0.133 should be added to both examples when generating the true CERFs. This resolves the lack of coverage in the second example.
```r
library(GPCERF)
library(SuperLearner)
library(ggplot2)
set.seed(32236)

N <- 500
use_nn <- FALSE

sim_data <- data.frame(Y = rep(NA_real_, N), treat = rnorm(N), cf1 = rnorm(N), cf2 = rnorm(N))
sim_data$Y <- rnorm(N, mean = cos(2 * sim_data$cf1))

# exposure values
q1 <- stats::quantile(sim_data$treat, 0.05)
q2 <- stats::quantile(sim_data$treat, 0.95)

true_df <- data.frame(w = seq(q1, q2, length.out = 100))
true_df$y = 0.133

n_core <- 12

m_xgboost <- function(nthread = n_core, ...) {
  SuperLearner::SL.xgboost(nthread = nthread, ...)
}

m_ranger <- function(num.threads = n_core, ...){
  SuperLearner::SL.ranger(num.threads = num.threads, ...)
}

# Estimate GPS function
gps_m <- GPCERF::estimate_gps(cov_mt = sim_data[,-(1:2)],
                      w_all = sim_data$treat,
                      sl_lib = c("SL.mean", "SL.glm", "SL.gam", "SL.glm.interaction",
                                 "SL.earth", "m_xgboost", "m_ranger"),
                      dnorm_log = TRUE)

w_all <- seq(q1, q2, length.out = 100)

# test = gam( Y ~ s(treat), data = sim_data)
# test_pred = predict(test, data.frame(treat = w_all), se.fit = T)
# plot(w_all, test_pred$fit, type = "l", ylim = c(-1.5,1.5))
# lines(w_all, test_pred$fit + test_pred$se.fit*1.96, lty = 2)
# lines(w_all, test_pred$fit - test_pred$se.fit*1.96, lty = 2)
# lines(w_all, true_df$y, col = "red")

params_lst <- list(alpha = 10 ^ seq(-2, 1/2, length.out = 10),
                   beta = 10 ^ seq(-2, 1/2, length.out = 10),
                   g_sigma = c(0.1, 0.5, 1, 5, 10),
                   tune_app = "all")

if(use_nn) {
  params_lst$n_neighbor <- 50
  params_lst$block_size <-  1e3
  cerf_gp_obj <- estimate_cerf_nngp(sim_data,
                                    w_all,
                                    gps_m,
                                    params = params_lst,
                                    outcome_col = "Y",
                                    treatment_col = "treat",
                                    covariates_col = c("cf1", "cf2"),
                                    nthread = 10)
} else {
  cerf_gp_obj <- estimate_cerf_gp(sim_data,
                                  w_all,
                                  gps_m,
                                  params = params_lst,
                                  outcome_col = "Y",
                                  treatment_col = "treat",
                                  covariates_col = c("cf1", "cf2"),
                                  nthread = n_core)
}

plot_fun(cerf_gp_obj) +
  geom_line(data = true_df,
            aes(x = w, y = y, color = "True"), linewidth = 1.5)
```

We are aware that this modification would not improve the coverage of the first example. We would like to address this remaining issue of overconfidence by noting that in the first example the treatment  is independent of the two covariates  and . In other words, the data in the first example come from a randomized trial, rather than an observation study, which is the main target of our approach. In this case, the true GPS function  is independent of  and the estimated GPS may be highly unstable, especially when boosting based approaches (e.g., xgboost) are included in the ensemble. Since the hyper-parameter tuning procedure relies heavily on the estimated GPS, the selected values of hyper-parameters can be unreasonable. In fact, with the hyper-parameter grid provided in the example, our approach tends to select very small , which is equivalent to setting the signal-to-noise ratio of the model low, and as a result, the estimated curve is much flatter than it should be with a confidence band that barely covers the truth.
We propose to resolve the issue in example 1, and all scenarios where  is (nearly) independent of the list of available covariates, by tweaking the hyper-parameter grid (see also response to the comment 7). Since the GPS should not provide additional information to the estimation process, we fix the hyper-parameter for estimated GPS at a large value to essentially eliminate its influence on the final estimation. We then make sure the information for inference is mainly from the observed level of exposure  by setting the candidate range of it to be small, for example, between  and . We also focus on relatively large values of , reflecting a prior belief that there is some relationship between  and the outcome . The updated grid is
