### 2023-10-05 NSAPH-Software/GPCERF#72 @boyuren158
We first summarize the interpretation of the three hyper-parameters and the impact of their values on the estimation to motivate the various choices of hyper-parameter grid we propose. The scaling parameters $\alpha$ and $\beta$ determine how much information the estimation will draw from the two coordinates, GPS score and exposure level. A large scaling parameter suggests that varying the corresponding coordinate is only associated with a small change in the outcome, that is, this coordinate does not contribute too much to the variation of the outcome. The noise-to-ratio parameter $\gamma$ encodes how different the observed data is from pure noise. A large $\gamma$ indicates strong associations between the outcome and the coordinates of the GP while small $\gamma$ suggests the observed outcome is likely to be drawn from a random process that is independent of the coordinates.

In the setting of observational studies and under the no unobserved confounding assumption (note this is the setting for which GPCERF is specifically designed for), both the exposure level and the GPS score encode important information for the estimation of CERF, as a result, the range of their scaling parameters should be comparable and the covariate balance will determine which coordinate is more important. The range should also cover both ends of the importance from extremely important to nearly irrelevant. We choose to achieve this by considering the range on the log-10 scale with equally spaced candidate values. The specification of range of $\gamma$ also follows the same rationale when the prior belief about the strength of causal effect of the exposure is weak. One example grid is the one used in the vignette.
```r
params_lst <- list(alpha = 10 ^ seq(-2, 2, length.out = 10),
                   beta = 10 ^ seq(-2, 2, length.out = 10),
                   g_sigma = c(0.1, 1, 10),
                   tune_app = "all")
```

However, as the reviewer correctly pointed out, the observed data might be generated from randomized trials where the exposure is completely independent of any available covariates. In this case, the estimated GPS score is in fact a noisy version of the exposure level (since the mean of the exposure level is not a function of the covariates) and it does not provide any additional information beyond the exposure itself. In this case, we have strong prior belief that the importance of the GPS score is low and thus a large value of $\alpha$ can be specified without any additional tuning of it. The range of $\beta$ should focus on smaller values to capture the information carried by the exposure. The specification of $\gamma$ should rely on some qualitative characterization, such as simple visualization, of the relationship between the outcome and the exposure. The reason is that under randomized design, the CERF can be estimated by regressing the outcome on the exposure directly with a flexible model. If the scatterplot of the outcome vs. the exposure exhibits clear trend, we should specify a large $\gamma$ to reflect the prior belief of a strong causal effect of the exposure. An example of this hyper-parameter grid specification can be found in the response to Comment 6.

In reality, we might not know whether the study is a randomized trial or not. We can then compute the absolute correlations between the exposure and each of the covariates to detect potential deviation from randomization. When all the absolute correlation is small than a pre-specified threshold, for instance 0.05, we can employ the specification for hyper-parameter grid as described above for randomized trials.
