### 2023-10-05 #5465 @Naeemkh
Hello @martinmodrak,

Thank you for reviewing both the manuscript and the code. We've addressed all the issues you highlighted, and these changes have improved both the manuscript and the code. Could you please take a moment to review these updates? If you find the modifications satisfactory, we kindly ask you to close the issues. If not, please don't hesitate to leave a comment for further discussion.

https://github.com/NSAPH-Software/GPCERF/issues/66
https://github.com/NSAPH-Software/GPCERF/issues/67
https://github.com/NSAPH-Software/GPCERF/issues/68
https://github.com/NSAPH-Software/GPCERF/issues/69
https://github.com/NSAPH-Software/GPCERF/issues/70
https://github.com/NSAPH-Software/GPCERF/issues/71
https://github.com/NSAPH-Software/GPCERF/issues/72

Thank you,
The authors
