### 2023-11-20 #5465 @spholmes
Hello @martinmodrak , would you be OK in deciding that the authors response to issue [#71 ](https://github.com/NSAPH-Software/GPCERF/issues/71) is acceptable and you could close it now? That would enable us to move on to finishing the reviewing process. Many Thanks, Susan
