### 2023-11-20 NSAPH-Software/GPCERF#72 @martinmodrak
This is a nice discussion of some of the issues, thanks. I'd just expect it to be covered in the documentation, not in GitHub issues. Note that I am not asking this specifically related to the problems I encountered in the other issues, I just think users should be able to learn this information somewhere. If it is in the docs and I overlooked than sorry.
