### 2023-11-20 NSAPH-Software/GPCERF#67 @martinmodrak
Thanks, this clarifies everything.
