### 2023-11-20 NSAPH-Software/GPCERF#70 @martinmodrak
There appears to be no change to the docs website or the readme (presumably because you are working on a separate JOSS branch ? But the README is identical there as well ... ) making it a bit hard for me to see and test the changes made... Could you update the docs or show me some examples on how the new interface looks like?
