### 2023-11-20 NSAPH-Software/GPCERF#72 @Naeemkh
Hi @martinmodrak, thanks for your comments. The discussion has been added to the paper in the following commit (which is included in the updated JOSS branch): https://github.com/NSAPH-Software/GPCERF/commit/b55c1844dfe329f995c47040f15a3fc778ebc173
