### 2023-11-20 NSAPH-Software/GPCERF#72 @martinmodrak
Nice, although I think it would be much more useful in the docs for `estimate_cerf_gp`, than in the paper (i.e. at https://nsaph-software.github.io/GPCERF/reference/estimate_cerf_gp.html) - seems a bit too low level of detail for paper while most users will likely not expect to find that info in the paper, while they may look at the docs.
