### 2023-11-21 NSAPH-Software/GPCERF#72 @Naeemkh
@martinmodrak, I agree. We added "A Note on Choosing Hyperparameters" section to the docs. Here is the link: https://nsaph-software.github.io/GPCERF/articles/A-Note-on-Choosing-Hyperparameters.html
