### 2023-11-21 NSAPH-Software/GPCERF#70 @Naeemkh
Hi @martinmodrak,
Thanks for the comment. Much of the work on this issue centered around addressing the background bug you accurately identified, which occurred when selecting a single covariate in the study. Regarding the interface, we have eliminated the specific ordering of data and now gather it according to each column name. In forthcoming major updates, we plan to incorporate a formula for data handling. We updated the documentation. In the readme file example, the following input values have been added.

```
 outcome_col = "Y",
 treatment_col = "treat",
 covariates_col = paste0("cf", seq(1,6)),
```
