### 2023-11-27 NSAPH-Software/GPCERF#72 @martinmodrak
Nice!
