### 2023-12-17 #5465 @andrewherren
Hi all, thanks for the push here. One clarifying question: is my assessment of the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) correct that my role as a reviewer is simply to verify that the software works as promised and follows open source best practices? Without any need to comment / weigh in on the GPCERF method itself?
