### 2023-12-17 #5465 @spholmes
Hi @andrewherren , thanks for the question, yes the software is the focus of the review, the actually statistical methodology does not have to be commented on.
