### 2023-12-17 #5465 @martinmodrak
Hi, I didn't have time to fully complete the resolution to https://github.com/NSAPH-Software/GPCERF/issues/71 but I don't want to make the authors wait more - since at least part of the problem was my misunderstanding, I'll grant that the rest is too. I have no other outstanding issue with the submission.
