### 2023-12-17 #5465 @andrewherren
I understand that I am supposed to conduct my review via Github issues in the original repo, but I have a number of comments on the JOSS paper draft itself, which is not included in the GPCERF repo, so I will submit them here unless otherwise instructed.

1. The statement of need makes a claim about GPCERF's performance ("Since standard GPs are infamous for their scalability issues - particularly due to operations involving the inversion of covariance matrices - we adopt a nearest-neighbor GP (nnGP) prior to ensure computationally efficient inference of the CERF in large-scale datasets") and then does not elaborate or validate the claim. Specifically, is the claim about the runtime of GPCERF under the nnGP prior (i.e. something faster than O(n^3))? As a reader, I am left to trust this claim without any evidence.
2. Similarly, the statement of need claims that resampling-based approaches, such as CausalGPS which the authors cite, are computationally burdensome compared to GPCERF without any comparison of runtimes of the two methods.
3. There are several typos in the [article proof](https://github.com/openjournals/joss-papers/blob/joss.05465/joss.05465/10.21105.joss.05465.pdf). For example "covaraite balance" and "pseduo-population" on page 2.
