### 2023-12-18 #5465 @andrewherren
I've just found a version of the [paper](https://github.com/NSAPH-Software/GPCERF/blob/JOSS/paper/paper.md) in the `JOSS` branch of the `GPCERF` repo, so I will continue my review via Github issues.
