### 2023-12-18 NSAPH-Software/GPCERF#78 @andrewherren
Statement of Need - Confirm / Validate Performance Claims
One of the checklist items for the JOSS review asks reviewers "if there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)"

In the case of [the 12/17/23 version of the JOSS draft](https://github.com/NSAPH-Software/GPCERF/blob/1e9ea19a4ba2752882a9747c84543c9bd392a346/paper/paper.md), there are several points made about the performance of GPCERF that are not backed up by clear evidence. Specifically:

1. The computational performance of nnGP is contrasted with that of a traditional GP (`Since standard GPs are infamous for their scalability issues - particularly due to operations involving the inversion of covariance matrices - we adopt a nearest-neighbor GP (nnGP) prior to ensure computationally efficient inference of the CERF in large-scale datasets`). However, the paper does not elaborate on or validate these claims. Specifically, is the claim that the nnGP prior reduces the time complexity from `O(n^3)`? If not (and it's completely fine if not!), what is the specific performance advantage of the nnGP? One way to illustrate this point might be a graph comparing runtimes for nnGP and standard GP for different sample sizes. But as it stands, the paper is just asserting that the nnGP prior is addressing a specific need because it's faster, without any evidence of being faster.
2. More or less as above, the statement of need claims that GPCERF is computationally advantageous compared to resampling-based approaches without any specific evidence of a performance difference.
