### 2023-12-19 #5465 @andrewherren
Thank you, @Naeemkh. Am I looking at the correct update [here](https://github.com/NSAPH-Software/GPCERF/blob/JOSS/paper/paper.md)?

I don't see an update on issue 78 and, regarding issue 79, I think it behooves you to explain why `GPCERF` should only be compared to Bayesian nonparametric methods. In my view, readers are generally better served by always knowing more about what's available. The reviewer guidelines ask if the statement of need "clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work." I don't quite understand why other work being frequentist means it is entirely unrelated to solving the problem that GPCERF is solving.

Happy to defer to the editor's judgment on this point, I just feel the paper as currently written presents a slightly constrained view of the open source software ecosystem for this problem space.
