### 2023-12-19 #5465 @Naeemkh
> Thank you. Same points as above still hold, regarding issues 78 and 79.

Hi @andrewherren,

I might missing something. You highlighted two key points in [issue 78](https://github.com/NSAPH-Software/GPCERF/issues/78#issue-2045744832):

Our paper originally claimed, "The computational performance of nnGP is contrasted with that of a traditional GP." You suggested demonstrating this through a graphical comparison of runtimes for nnGP and standard GP across different sample sizes. Previously, the paper asserted the nnGP prior's speed advantage without direct evidence.

- In response, we have included Figure 3 and Figure 4 to address this.

You pointed out that our statement of need claimed GPCERF's computational advantage over resampling-based approaches without providing specific evidence of a performance difference.

- We have removed this claim, as exploring and providing evidence for performance differences between GPCERF and other software is beyond the scope of our work and the Journal of Open Source Software (JOSS).

Please let us know if there are any further aspects we need to address.
