### 2023-12-20 #5465 @andrewherren
I also think that

> Due to the property of tree-based approaches, the resulting CERF is usually not smooth.

is a bit of an exaggerated claim about tree ensembles. A well-trained ensemble of 200 trees, while not satisfying the mathematical definition of continuity, will often resemble a smooth approximation, if the underlying signal in a dataset is smooth.

But also, the claim sort of seems to assume that ERFs should be _a priori_ expected to be smooth. Is there any reason readers should be led to believe that smooth ERFs are desirable? It seems entirely plausible to imagine a dose-response relationship which is more of a step function (i.e. doses in a given narrow range all have the same effect), which is something that a tree ensemble can typically estimate well but a squared-exponential GP will not estimate as easily.

I don't understand the necessity of dismissing tree-based approaches like this. Why not just present it as "here is another way that open source authors have thought to estimate CERFs"?
