### 2023-12-20 #5465 @Naeemkh
Hi @andrewherren,

Thank you for the comment. First, we would like to clarify that we consider a function smooth if it is everywhere differentiable. It is well-known that a feature of ensembles of decision trees is their lack of smoothness (see for example, Linero and Yang, 2018; Hill et al, 2020). Indeed, with a large number of trees, BART resembles nowhere-differentiable continuous function and Linero and Yang (2018) showed that when the number of trees goes to infinity, BART prior converges to a nowhere-differentiable Gaussian process. This does not imply that BART cannot approximate a smooth function well. However, methods that can adapt to higher level of smoothness typically will outperform BART in terms of, for example, posterior contraction rate, if the underlying regression function is smooth. The success of SoftBart (Linero, 2022) clearly illustrates this point. Therefore, we believe that our statement is not an exaggeration—it only reiterates a mathematical property of BART.

We also want to clarify that we have no intention to dismiss tree-based approaches by mentioning their lack of smoothness. Indeed, as you have pointed out, if the underlying CERF behaves like a step-function, BART will likely perform well. However, it is equally plausible to assume that the true CERF is smooth, if the researchers believe even a small increase in the exposure would lead to changes in the outcome of interest. In this case, a CERF estimation procedure that can adapt to a desired level of smoothness is clearly preferable to a BART-based approach. We highlight the lack of smoothness of BART to inform readers that it might not be the optimal choice when there is substantial prior evidence indicating a high level of smoothness of the underlying CERF. We were meant to mention that this limitation can be resolved by replacing BART with alternatives such as SoftBart. But since causaldrf does not offer such an option, we chose to not discuss this point further. If you think it is helpful, we could add this note to the Need of Statement. Finally, throughout the paper, we refrained from delving into the algorithmic intricacies of the various options for CERF estimation, as such a discussion is more appropriate for a methodological paper than a software paper.

Thanks,
The authors
