### 2023-12-20 #5465 @andrewherren
I certainly would not contend that tree ensembles satisfy the mathematical definition of smoothness, apologies if I gave that impression.

My hang-up is that the article, as currently written, doesn't make it clear that GPCERF is **_explicitly_** regularizing toward smooth CERFs. It's fine to present it as "if you think your CERF is smooth, GPCERF may be preferable because of its regularization scheme, and if you think it is discontinuous, BART-based methods may be preferable because their regularization prior makes it easier to estimate discontinuous functions." Right now the article just says that BART is non-smooth, without making it clear that GPCERF is explicitly smooth in a way that makes recovering discontinuous DGPs considerably more difficult. I think the statement of need should be written with a more acute presentation of the tradeoffs that abound in nonparametric estimation.

On the point about SoftBart, I don't feel strongly about mentioning it in the statement of need, since as you mentioned there are no available open source implementations of SoftBart for CERF estimation and this is a software paper.

Could you clarify a bit more what you are referring to regarding:

> Finally, throughout the paper, we refrained from delving into the algorithmic intricacies of the various options for CERF estimation, as such a discussion is more appropriate for a methodological paper than a software paper.

I agree that it's not necessary to dive into algorithmic intricacies in citing other packages that estimate the CERF.
