### 2023-12-21 #5465 @Naeemkh
Hi @andrewherren,

Thank you very much for the clarification and all the great suggestions on the statement of need. We have revised the section to 1) mention explicitly that GPCERF can accommodate different levels of smoothness of the underlying causal ERF by using specifically designed covariance functions; 2) add references to some of the commonly used R packages for estimating causal ERF that are built on frequentist frameworks; and 3) add a short note to point the readers to Figure 3 and 4 for evidence of the computational scalability of nnGP in comparison to standard GP.

Regarding our previous comment about not delving into algorithmic intricacies, we were mainly referring to the consideration of including SoftBart in the discussion about casualdrf. We apologize for the confusion.

Best regards,
The authors
