### 2024-01-23 #5465 @andrewherren
A few comments that remain on my checklist:

> Functionality: Have the functional claims of the software been confirmed?

If I'm missing the scripts that produced Figures 3 and 4 in the current version of the paper, feel free to point me in the right direction. I want to make sure the functional claims about scalability of the package are reproducible.

> A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

I think in the paper's current state, the package lacks a clear motivation. Earlier in this review, I noted that the paper asserted that `GPCERF` was more scalable than `CausalGPS` without any plots / simulation studies / other evidence of that comparison. As far as I can tell, the response was to remove comments about the relative scalability of the two methods. That is fine, but then why are users better served by `GPCERF` over `CausalGPS`? Similarly, the BART-based CERF estimator in `causaldrf` is cited alongside a caveat that "the resulting CERF is usually not smooth due to the use of regression trees." The statement is true at face value, but it appears to blend two points:

1. The mathematical fact that regression trees (and ensembles of regression trees) are discontinuous
2. A notion that, because tree ensembles fail the mathematical definition of continuity, they must underperform continuous GP-based estimators when the underlying CERF is continuous

2 seems to be the actual reason why users would prefer `GPCERF` over discontinuous estimators such as the BART based estimator in `causaldrf`. Without backing up this claim, I don't see why it is a problem that "the resulting CERF is usually not smooth due to the use of regression trees," or why users should worry about this.

If the reason to prefer this package is that it's faster than competitors, then that should be demonstrated (a quick plot of a simulation study would suffice). If the reason to prefer this package is that its continuous GP-based estimator has favorable estimation properties in the case of a continuous CERF, then that should be demonstrated.
