### 2024-01-23 #5465 @andrewherren
Once I have been able to run the scripts producing Figures 3 and 4, I will defer to @spholmes on the second half of my comment above.

I feel the paper's statement of need is not entirely well-motivated, but if I am somehow misreading that part of the checklist, then I won't insist on resolving that comment.
