### 2024-01-23 #5465 @spholmes
Thanks @drewherren, we will wait for the scripts for Figures 3 and 4 and then I will do a complete re-read and check the need-defining section, in particular the continuity motivation.
