### 2024-01-23 #5465 @Naeemkh
@andrewherren and @spholmes , Thank you very much for your input.

We have uploaded the code necessary to replicate the figures from the paper in the following folder:
- https://github.com/NSAPH-Software/GPCERF/tree/JOSS/functional_tests/paper_runs

Detailed steps to initialize a container and execute the code can be found in the accompanying [readme](https://github.com/NSAPH-Software/GPCERF/blob/JOSS/functional_tests/paper_runs/README.md) file.

Regarding your second point, we have updated the description of SON as follows:
causaldrf [@causaldrf_R] uses Bayesian Additive Regression Trees (BART) for flexible causal ERF estimation _and it is particularly useful when the underlying ERF is discontinuous or resembles step functions_.

As for comparing our package with others mentioned in this study, such an analysis would necessitate a thorough review of all the assumptions recommended by these packages, which falls outside the scope of our current paper.

Best regards,
The authors
