### 2024-01-24 #5465 @andrewherren
Thank you, I will run those examples!

On the second point, we may be speaking past each other a bit. The point I am making is perhaps better illustrated with a table of two different estimation scenarios

| Underlying Data Generating Process (DGP) | Estimators |
| --- | --- |
| Continuous | Continuous (e.g. GP) or Discontinuous (e.g. tree ensemble) |
| Discrete | Continuous (e.g. GP) or Discontinuous (e.g. tree ensemble) |

I haven't come across any evidence that continuous estimators always (or even generally) outperform discrete estimators on continuous DGPs, nor vice versa.

The suggestion in the paper, that GP-based estimators are best when the underlying CERF is continuous and tree ensembles are best when the underlying CERF is discontinuous, strikes me as unfounded. If there's concrete evidence that supports the claim, that would be worth including in the paper. But as an assertion, I'm not sure it holds up.
