### 2024-01-29 #5465 @Naeemkh
Hi @andrewherren,

We agree with your comment. Thus, we removed the statement about the strengths and limitations of BART from the statement of need. GPCERF aims to introduce a new approach for CERF estimation using Bayesian nonparametrics, which has not been used extensively in the causal inference of continuous exposure.

Thanks,
The authors
