### 2024-02-22 #5465 @spholmes
Hello @andrewherren and @Naeemkh
I am stepping in here, if I don't hear any objections in the next two days, I will be going ahead with the editorial process to accept.
In the meantime, @Naeemkh if you could recheck some of the references, as some packages also seem to have publications attached to them that you could cite and there are a few places in the journal articles where the capitalization should be conserved using {} ({Gaussian} for instance.
Thanks, Susan
