### 2024-02-23 #5465 @Naeemkh
Hello @spholmes,

Thank you. I've included the method papers for the R packages mentioned, and I've also double-checked the formatting. Within the next few hours, I will compile the document here and perform another round of checks.

Best regards,
Naeem
