### 2024-02-23 #5465 @andrewherren
Hi all,

Thank you for the push here. Some of the replication scripts take a long time to run, so I will defer to the authors that the figures in the paper are good to go. If everyone else is good with the Statement of Need section, then all is good to go on my end and I'll check off everything in the checklist.

Separately, some feedback on the review process is that it was pretty unclear what my role here was supposed to be. I suspect based on the discussion in this thread that my role was intended to be narrowly focused on making sure the package works and is documented / tested (as opposed to asking for more motivation on why a user would prefer `GPCERF` over its alternatives)? Might be worth clarifying that for future referees.

Best,
Drew
