### 2024-02-23 #5465 @spholmes
Dear @andrewherren,
No worries, you did exactly what we wanted, reviewing does not mean just testing: the needs section is as important as the rest and an honest review of the landscape of choices is very important to the article, so we are very thankful you paid attention to that. We provide the checklist for you to go over each of the points as easily as possible but all the items are important in the review.
I hope that as machines improve the computations are not as slow and it is good that this is noted "in a public space".
Many thanks for all your input here.
Susan
