### 2024-02-26 #5465 @Naeemkh
Hi @Drew,

As the wall clock time logs show (`functional_tests/paper_runs`), replicating scaling examples will take around 17 hours. You can select a couple of instances and run them (e.g., seed value: 19259, data size: 3000, modeling approach: GP), instead of running the entire instances. Running all examples by the reviewers might not be feasible in all cases. For example, if I submit another paper and use up to 100,000 CPU cores in scaling tests, we cannot expect the reviewers to reproduce the results.

Best regards,
Naeem
