### 2024-03-12 #5465 @spholmes
Then please move ahead with the zenodo DOI , I need this set up before I can move on to the final acceptance. You need to do as you did for CRE and create an archive (on [Zenodo](https://zenodo.org/)or [figshare](https://figshare.com/), or other) and post the archive DOI in the REVIEW issue. Then I will ask add the archive to the issue with the correct DOI for the zenodo archive.
