### 2024-03-12 #5465 @Naeemkh
> Hello @spholmes,
>
> We released the package on CRAN: https://cran.r-project.org/web/packages/GPCERF/index.html (v0.2.3) Also we archived the release on Zenodo: https://zenodo.org/records/10757333 DOI: 10.5281/zenodo.10757333
>
> Please let us know if we need to take any other actions.
>
> Thanks, The authors

@spholmes , please see my previous comment. I am not sure if I need to take any other actions, other than mentioned in the comment. Please let me know. Thanks.
